Separated some helper/parsers method to helper.h
